% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fulcrum_photo}
\alias{fulcrum_photo}
\title{Get Fulcrum photo}
\usage{
fulcrum_photo(
  photo_id,
  api_key = get_api_key(),
  base_url = "https://api.fulcrumapp.com/api/v2/photos/"
)
}
\arguments{
\item{photo_id}{Photo ID from Fulcrum}

\item{api_key}{Fulcrum authentication key}

\item{base_url}{Base URL for Fulcrum Photo API}
}
\value{
If request is successful (i.e. the request was successfully performed
and a response with HTTP status code <400 was recieved), an HTTP response; otherwise throws an error.
}
\description{
Submit GET request to Fulcrum Photo API to get individual photo metadata
}
