% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fulcrum_query}
\alias{fulcrum_query}
\title{Query Fulcrum API}
\usage{
fulcrum_query(
  query_string,
  api_key = get_api_key(),
  base_url = "https://api.fulcrumapp.com/api/v2/query/"
)
}
\arguments{
\item{query_string}{SQL statement as string}

\item{api_key}{Fulcrum authentication token stored as environment variable}

\item{base_url}{Base URL for Fulcrum Query API}
}
\description{
Submit GET request to Fulcrum Query API based on SQL query string
}
\details{
Returns parsed JSON body from response to Fulcrum query.
}
