% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fulcrum_table}
\alias{fulcrum_table}
\title{Get Fulcrum table}
\usage{
fulcrum_table(table_name, api_key = get_api_key(), col_types = NULL)
}
\arguments{
\item{table_name}{Name of Fulcrum table}

\item{api_key}{Fulcrum authentication key}

\item{col_types}{One of NULL, a cols() specification, or a string. See vignette("readr") for more details.}
}
\value{
A data frame for specified Fulcrum table
}
\description{
Get Fulcrum table with specified \code{table_name}
}
