% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fulcrum_all_tables}
\alias{fulcrum_all_tables}
\title{Get Fulcrum tables}
\usage{
fulcrum_all_tables(api_key = get_api_key())
}
\arguments{
\item{api_key}{Fulcrum authentication token stored as environment variable}
}
\description{
Get all available Fulcrum tables
}
\details{
Returns data frame with name, type, etc. of all available Fulcrum tables
}
