% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fulcrum_query}
\alias{fulcrum_query}
\title{Query Fulcrum API}
\usage{
fulcrum_query(
  query_string,
  api_key = get_api_key(),
  base_url = "https://api.fulcrumapp.com/api/v2/query/"
)
}
\arguments{
\item{query_string}{SQL statement as string}

\item{api_key}{Fulcrum authentication key}

\item{base_url}{Base URL for Fulcrum Query API}
}
\value{
If request is successful (i.e. the request was successfully performed
and a response with HTTP status code <400 was recieved), an HTTP response; otherwise throws an error.
}
\description{
Submit GET request to Fulcrum Query API based on SQL query string
}
