% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{set_api_key}
\alias{set_api_key}
\alias{get_api_key}
\title{Set or get Fulcrum API key}
\usage{
set_api_key(key = NULL)

get_api_key()
}
\description{
Set or get Fulcrum API key as environmental variable in .Renviron
}
